Receiver/requirements.txt
connexion==2.9.0
requests==2.26.0
PyYAML==5.4.1
pykafka==2.8.0